+++ .idea/compiler.xml
@@ -6,7 +6,9 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="MVCloud" />
+        <module name="Server" />
+        <module name="Common" />
+        <module name="Client" />
       </profile>
     </annotationProcessing>
   </component>
