+++ .idea/dictionaries/Maxim.xml
@@ -1,0 +1,7 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="Maxim">
+    <words>
+      <w>portu</w>
+    </words>
+  </dictionary>
+</component>
