+++ Client/src/main/java/Common.java
@@ -1,0 +1,10 @@
+import java.net.Socket;
+import java.nio.file.Path;
+
+public abstract class Common {
+
+    public static Socket socket;
+    public static AuthList_lite user;
+    public static Path path;
+
+}
