+++ Common/src/main/java/AuthList_lite.java
@@ -1,0 +1,27 @@
+import java.io.Serializable;
+
+public class AuthList_lite implements Serializable {
+    private String name;
+    private String pass;
+
+    public AuthList_lite(String name, String pass) {
+        this.name = name;
+        this.pass = pass;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getPass() {
+        return pass;
+    }
+
+    public void setPass(String pass) {
+        this.pass = pass;
+    }
+}
