+++ Common/src/main/java/UserAuth.java
@@ -1,0 +1,22 @@
+import java.io.Serializable;
+
+public class UserAuth implements Serializable {
+    private String name;
+    private String pass;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getPass() {
+        return pass;
+    }
+
+    public void setPass(String pass) {
+        this.pass = pass;
+    }
+}
