+++ Server/src/main/java/SController.java
@@ -1,0 +1,64 @@
+import io.netty.bootstrap.ServerBootstrap;
+import io.netty.channel.ChannelFuture;
+import io.netty.channel.ChannelInitializer;
+import io.netty.channel.EventLoopGroup;
+import io.netty.channel.nio.NioEventLoopGroup;
+import io.netty.channel.socket.SocketChannel;
+import io.netty.channel.socket.nio.NioServerSocketChannel;
+import io.netty.handler.codec.serialization.ClassResolvers;
+import io.netty.handler.codec.serialization.ObjectDecoder;
+import io.netty.handler.codec.serialization.ObjectEncoder;
+import javafx.fxml.Initializable;
+import javafx.scene.control.Button;
+import javafx.scene.control.TextField;
+
+import java.net.URL;
+import java.util.ResourceBundle;
+
+public class SController implements Runnable, Initializable {
+
+    public TextField address;
+    public TextField portu;
+    public Button startu;
+
+    @Override
+    public void initialize(URL location, ResourceBundle resources)  {
+        System.out.println("Init");
+    }
+
+    @Override
+    public void run() {
+        EventLoopGroup auth = new NioEventLoopGroup(1);
+        EventLoopGroup worker = new NioEventLoopGroup();
+        try {
+            ServerBootstrap bootstrap = new ServerBootstrap();
+            bootstrap.group(auth, worker)
+                    .channel(NioServerSocketChannel.class)
+                    .childHandler(new ChannelInitializer<SocketChannel>() {
+                        @Override
+                        protected void initChannel(SocketChannel socketChannel) throws Exception {
+                            socketChannel.pipeline().addLast(
+                                    new ObjectEncoder(),
+                                    new ObjectDecoder(ClassResolvers.cacheDisabled(null)),
+                                    new ClientHandlerReader());
+                        }
+                    });
+            ChannelFuture future = bootstrap.bind(ServerData.adress, Integer.parseInt(ServerData.port)).sync();
+            System.out.println("server started");
+            future.channel().closeFuture().sync(); // block
+        } catch (Exception e) {
+            e.printStackTrace();
+        } finally {
+            auth.shutdownGracefully();
+            worker.shutdownGracefully();
+        }
+    }
+
+    public void startServer(javafx.event.ActionEvent actionEvent){
+        ServerData.adress = address.getText();
+        ServerData.port = portu.getText();
+        new Thread(new Server()).start();
+    }
+
+
+}
