+++ Server/src/main/java/ServerData.java
@@ -1,0 +1,6 @@
+public abstract class ServerData {
+
+    public static String adress;
+    public static String port;
+
+}
