+++ Server/src/main/resources/ServerFiles/Max/ozon.txt
