+++ pom.xml
@@ -6,6 +6,7 @@
 
     <groupId>org.example</groupId>
     <artifactId>MVCloud</artifactId>
+    <packaging>pom</packaging>
     <version>1.0-SNAPSHOT</version>
 
 
@@ -22,6 +23,11 @@
             </plugin>
         </plugins>
     </build>
+    <modules>
+        <module>Client</module>
+        <module>Server</module>
+        <module>Common</module>
+    </modules>
 
     <dependencies>
         <dependency>
