--- src/main/java/Client/Common.java
@@ -1,10 +1,0 @@
-package Client;
-
-import java.net.Socket;
-
-public abstract class Common {
-
-    public static Socket socket;
-    public static AuthList_lite user;
-
-}
