--- src/main/resources/Authentification.fxml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
-<AnchorPane fx:id="Central" prefHeight="184.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/10.0.2-internal" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Authentification">
-   <children>
-      <Pane fx:id="parent" layoutY="8.0" prefHeight="175.0" prefWidth="588.0">
-         <children>
-            <PasswordField fx:id="pass" layoutX="14.0" layoutY="123.0" prefHeight="36.0" prefWidth="344.0" />
-            <TextField fx:id="name" layoutX="14.0" layoutY="62.0" prefHeight="36.0" prefWidth="344.0" />
-            <Button fx:id="auth" layoutX="418.0" layoutY="56.0" mnemonicParsing="false" onAction="#sendCommand" prefHeight="47.0" prefWidth="156.0" text="Authentification" />
-            <Label layoutX="17.0" layoutY="38.0" prefHeight="18.0" prefWidth="96.0" text="Name" />
-            <Label layoutX="14.0" layoutY="103.0" prefHeight="18.0" prefWidth="84.0" text="Password" />
-            <Label layoutX="190.0" layoutY="4.0" prefHeight="29.0" prefWidth="208.0" text="              AUTHENTIFICATION" />
-         </children></Pane>
-   </children>
-</AnchorPane>
