--- src/main/resources/Max/UserFile.txt
@@ -1,7 +1,0 @@
-Проверка файла. Проверка фыавфыапвфыафвы афвы а
-ыв афыва фвы
-
-ыфвафвыплукплдлукодльлсчмтлвТЫав
-ываоуклатлт ЬБВм
-опривет, конечно же я пределаяся
-проверяем, насколько корректно передаётся файл
