--- src/main/resources/UsersData/UData.txt
@@ -1,2 +1,0 @@
-Max Pass
-Ksu Pass
